+++ tt_metal/llrt/hal_1xx/wormhole/wh_hal_idle_eth.cpp
